Finished fleshing out the querying system
Pretty happy with this, should make it easy to evolve over time, if
the API ever gets updated. Might still want to add something in to
better handle an introduction of a v2, but that's for another day.

CategoryClient should be good at this point, just needs to be replicated
over to all of the other resource clients.