More API implementation and testing
Overhauled the testing so that tests will used the resource files
already being used by the Jackson tests. Will have a set of IT tests
included in for sanity to ensure that the real queries continue to work.

I think about 3 of the client classes are completely done at this point,
which means that they're fully documented and fully tested. Still a lot
more left to finish and clean up, but the end is in sight.