Adding in more clients and tests
Ground work should be done for Games to go in now. I had missed a whole
set of dependencies that Games need, so those resources are all now
supported, including tests.

A few more resources left to flush out with tests, and then it's just
the Games Client and tests.