Partially fixing issue #2
This only partly solves the problem identified in issue #2. There's
still no way for the user to easily make that request on their own
using the navigation links.

Not really sure how to solve this at this point, and pretty burned out
of working on this for now. Will address this the next time I get back
to this project.